requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
gTTS==2.5.4
idna==3.10
MouseInfo==0.1.3
numpy==2.2.2
opencv-python==4.11.0.86
pillow==11.1.0
playsound==1.3.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pytweening==1.2.0
requests==2.32.3
setuptools==75.8.0
urllib3==2.3.0
wheel==0.45.1
